refactor(svelte): split pointer pure unit tests by decision table
Move shared builders to pointer-fixtures and split the 687-line
pointer.test.ts into down/up, move/leave, and click/touch/constants suites.
Assertions preserved.
